Abstract time limit stuff into a couple of calls.
