+++ include/MemoryModel/PointerAnalysis.h
@@ -30,6 +30,9 @@
 #ifndef POINTERANALYSIS_H_
 #define POINTERANALYSIS_H_
 
+#include <unistd.h>
+#include <signal.h>
+
 #include "Graphs/PAG.h"
 #include "MemoryModel/ConditionalPT.h"
 #include "MemoryModel/AbstractPointsToDS.h"
@@ -269,6 +272,45 @@ public:
     //@}
 
 protected:
+    static void timeLimitReached(int signum)
+    {
+        std::cout.flush();
+        SVFUtil::outs().flush();
+        // TODO: output does not indicate which time limit is reached.
+        //       This can be better in the future.
+        SVFUtil::outs() << "WPA: time limit reached\n";
+        exit(101);
+    }
+
+    /// Starts an analysis timer. If timeLimit is 0, sets no timer.
+    /// If an alarm has already been set, does not set another.
+    /// Returns whether we set a timer or not.
+    inline bool startLimitTimer(unsigned timeLimit) const
+    {
+        if (timeLimit == 0) return false;
+
+        // If an alarm is already set, don't set another. That means this analysis
+        // is part of another which has a time limit.
+        unsigned remainingSeconds = alarm(0);
+        if (remainingSeconds != 0)
+        {
+            // Continue the previous alarm and move on.
+            alarm(remainingSeconds);
+            return false;
+        }
+
+        signal(SIGALRM, &timeLimitReached);
+        alarm(timeLimit);
+        return true;
+    }
+
+    /// Stops an analysis timer. limitTimerSet indicates whether the caller set the
+    /// timer or not (return value of startLimitTimer).
+    inline void stopLimitTimer(bool limitTimerSet)
+    {
+        if (limitTimerSet) alarm(0);
+    }
+
     /// Return all indirect callsites
     inline const CallSiteToFunPtrMap& getIndirectCallsites() const
     {
